+++ fabric-api-dev/src/main/java/net/fabricmc/fabric/mixin/dev/BlockBoxMixin.java
@@ -14,28 +14,22 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin;
+package net.fabricmc.fabric.mixin.dev;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import org.spongepowered.asm.mixin.Dynamic;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
-import net.minecraft.server.command.CommandManager;
+import net.minecraft.util.math.BlockBox;
 
 import net.fabricmc.fabric.FabricDev;
 
-@Mixin(CommandManager.class)
-public class CommandManagerMixin {
-	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.REGISTER_DEBUG_COMMANDS")
-	@ModifyExpressionValue(method = "<init>", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
+@Mixin(BlockBox.class)
+public class BlockBoxMixin {
+	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.THROW_ON_INVALID_BLOCK_BOXES")
+	@ModifyExpressionValue(method = "<init>(IIIIII)V", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
 	private static boolean fabric$mevIsDevelopmentForDevModule(boolean original) {
-		return original || FabricDev.REGISTER_DEBUG_COMMANDS;
-	}
-
-	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.ENABLE_COMMAND_EXCEPTION_LOGGING")
-	@ModifyExpressionValue(method = "execute", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
-	private static boolean fabric$mevIsDevelopmentForDevModule2(boolean original) {
-		return original || FabricDev.ENABLE_COMMAND_EXCEPTION_LOGGING;
+		return original || FabricDev.THROW_ON_INVALID_BLOCK_BOXES;
 	}
 }
