+++ fabric-api-dev/src/main/java/net/fabricmc/fabric/mixin/dev/EulaReaderMixin.java
@@ -14,22 +14,22 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin;
+package net.fabricmc.fabric.mixin.dev;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import org.spongepowered.asm.mixin.Dynamic;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
-import net.minecraft.structure.StructureTemplateManager;
+import net.minecraft.server.dedicated.EulaReader;
 
 import net.fabricmc.fabric.FabricDev;
 
-@Mixin(StructureTemplateManager.class)
-public class StructureTemplateManagerMixin {
-	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.ENABLE_LOADING_STRUCTURES_FROM_GAMETESTS")
-	@ModifyExpressionValue(method = "<init>", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
-	private static boolean fabric$mevIsDevelopmentForDevModule(boolean original) {
-		return original || FabricDev.ENABLE_LOADING_STRUCTURES_FROM_GAMETESTS;
+@Mixin(EulaReader.class)
+public class EulaReaderMixin {
+	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.ALWAYS_AGREE_TO_EULA")
+	@ModifyExpressionValue(method = {"<init>", "createEulaFile"}, at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
+	private boolean fabric$mevIsDevelopmentForDevModule(boolean original) {
+		return original || FabricDev.ALWAYS_AGREE_TO_EULA;
 	}
 }
