+++ fabric-api-dev/src/main/java/net/fabricmc/fabric/mixin/dev/StructureTemplateManagerMixin.java
@@ -14,22 +14,22 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin;
+package net.fabricmc.fabric.mixin.dev;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import org.spongepowered.asm.mixin.Dynamic;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 
-import net.minecraft.util.collection.Weight;
+import net.minecraft.structure.StructureTemplateManager;
 
 import net.fabricmc.fabric.FabricDev;
 
-@Mixin(Weight.class)
-public class WeightMixin {
-	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.ZERO_WEIGHT_WARNING")
-	@ModifyExpressionValue(method = "validate", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
+@Mixin(StructureTemplateManager.class)
+public class StructureTemplateManagerMixin {
+	@Dynamic("@ModifyExpressionValue's the FIELD GET of SharedConstants.isDevelopment to add a OR condition for FabricDev.ENABLE_LOADING_STRUCTURES_FROM_GAMETESTS")
+	@ModifyExpressionValue(method = "<init>", at = @At(value = "FIELD", target = "Lnet/minecraft/SharedConstants;isDevelopment:Z"))
 	private static boolean fabric$mevIsDevelopmentForDevModule(boolean original) {
-		return original || FabricDev.ZERO_WEIGHT_WARNING;
+		return original || FabricDev.ENABLE_LOADING_STRUCTURES_FROM_GAMETESTS;
 	}
 }
