+++ fabric-api-dev/src/main/resources/fabric-api-dev.mixins.json
@@ -3,17 +3,17 @@
   "package": "net.fabricmc.fabric.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
-    "ArgumentTypesMixin",
-    "BlockAndItemMixin",
-    "BlockBoxMixin",
-    "BootstrapMixin",
-    "ChunkMixin",
-    "CommandManagerMixin",
-    "EulaReaderMixin",
-    "MinecraftServerMixin",
-    "StructureTemplateManagerMixin",
-    "UtilMixin",
-    "WeightMixin"
+    "dev.ArgumentTypesMixin",
+    "dev.BlockAndItemMixin",
+    "dev.BlockBoxMixin",
+    "dev.BootstrapMixin",
+    "dev.ChunkMixin",
+    "dev.CommandManagerMixin",
+    "dev.EulaReaderMixin",
+    "dev.MinecraftServerMixin",
+    "dev.StructureTemplateManagerMixin",
+    "dev.UtilMixin",
+    "dev.WeightMixin"
   ],
   "injectors": {
     "defaultRequire": 1
